pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "thinkbox-search"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "qdrant-client",
    "pyyaml",
]

[project.scripts]
thinkbox-embed = "thinkbox_search.embed:main"
thinkbox-search = "thinkbox_search.search:main"
